🚚  refactor: 스토리지를 스토어로 이동
