+++ js/state/index.js
@@ -1,16 +1,15 @@
-import createStorage from '../utils/storage.js';
 
 const createStore = (initialState) => {
-  const storage = createStorage(initialState);
+  getAll() ?? setStorage(initialState);
 
   const listeners = [];
 
-  const getState = (key) => (key ? storage.get(key) : storage.getAll());
+  const getState = (key) => (key ? get(key) : getAll());
 
   const dispatch = (newState) => {
-    const prevState = storage.getAll();
+    const prevState = getAll();
 
-    storage.set({
+    setStorage({
       ...prevState,
       ...newState,
     });
@@ -28,3 +27,18 @@ const createStore = (initialState) => {
 };
 
 export default createStore;
+
+const VENDING_MACHINE = 'VENDING_MACHINE';
+
+const setStorage = (value, key = VENDING_MACHINE) =>
+    localStorage.setItem(key, JSON.stringify(value));
+
+const getAll = (key = VENDING_MACHINE) => {
+  const data = localStorage.getItem(key);
+  // if (!data) throw new Error('데이터가 없습니다.');
+
+  return JSON.parse(data);
+};
+const get = (stateKey) => getAll()[stateKey];
+
+const remove = (key = VENDING_MACHINE) => localStorage.removeItem(key);
